Mint terms

+++ test/runeid.test.ts
@@ -1,27 +1,28 @@
 import { RuneId } from '../src/runeid';
-import { u128 } from '../src/u128';
+import { u128 } from '../src/integer/u128';
+import { u64, u32 } from '../src/integer';
 
 describe('runeid', () => {
   test('delta', () => {
     let expected = [
-      new RuneId(4, 2),
-      new RuneId(1, 2),
-      new RuneId(1, 1),
-      new RuneId(3, 1),
-      new RuneId(2, 0),
+      new RuneId(u64(4), u32(2)),
+      new RuneId(u64(1), u32(2)),
+      new RuneId(u64(1), u32(1)),
+      new RuneId(u64(3), u32(1)),
+      new RuneId(u64(2), u32(0)),
     ];
 
     expected = RuneId.sort(expected);
 
     expect(expected).toEqual([
-      new RuneId(1, 1),
-      new RuneId(1, 2),
-      new RuneId(2, 0),
-      new RuneId(3, 1),
-      new RuneId(4, 2),
+      new RuneId(u64(1), u32(1)),
+      new RuneId(u64(1), u32(2)),
+      new RuneId(u64(2), u32(0)),
+      new RuneId(u64(3), u32(1)),
+      new RuneId(u64(4), u32(2)),
     ]);
 
-    let previous = new RuneId(0, 0);
+    let previous = new RuneId(u64(0), u32(0));
     const deltas: [u128, u128][] = [];
     for (const id of expected) {
       const delta = previous.delta(id).unwrap();
@@ -37,7 +38,7 @@ describe('runeid', () => {
       [1n, 2n],
     ]);
 
-    previous = new RuneId(0, 0);
+    previous = new RuneId(u64(0), u32(0));
     const actual: RuneId[] = [];
     for (const [block, tx] of deltas) {
       const next = previous.next(block, tx).unwrap();
@@ -49,7 +50,7 @@ describe('runeid', () => {
   });
 
   test('display', () => {
-    expect(RuneId.new(1, 2).unwrap().toString()).toBe('1:2');
+    expect(RuneId.new(u64(1), u32(2)).unwrap().toString()).toBe('1:2');
   });
 
   test('from string', () => {
@@ -58,6 +59,6 @@ describe('runeid', () => {
     expect(() => RuneId.fromString(':2')).toThrow();
     expect(() => RuneId.fromString('a:2')).toThrow();
     expect(() => RuneId.fromString('1:a')).toThrow();
-    expect(RuneId.fromString('1:2')).toEqual(new RuneId(1, 2));
+    expect(RuneId.fromString('1:2')).toEqual(new RuneId(u64(1), u32(2)));
   });
 });
